+++ java/jeeintegration/src/main/java/io/joynr/jeeintegration/DefaultJoynrRuntimeFactory.java
@@ -58,7 +58,6 @@ import io.joynr.capabilities.ParticipantIdKeyUtil;
 import io.joynr.exceptions.JoynrIllegalStateException;
 import io.joynr.jeeintegration.api.JeeIntegrationPropertyKeys;
 import io.joynr.jeeintegration.api.JoynrLocalDomain;
-import io.joynr.jeeintegration.api.JoynrMessagePersister;
 import io.joynr.jeeintegration.api.JoynrMqttClientIdProvider;
 import io.joynr.jeeintegration.api.JoynrProperties;
 import io.joynr.jeeintegration.api.JoynrRawMessagingPreprocessor;
@@ -69,8 +68,6 @@ import io.joynr.messaging.RawMessagingPreprocessor;
 import io.joynr.messaging.mqtt.MqttClientIdProvider;
 import io.joynr.statusmetrics.JoynrStatusMetrics;
 import io.joynr.statusmetrics.JoynrStatusMetricsReceiver;
-import io.joynr.messaging.persistence.MessagePersister;
-import io.joynr.messaging.persistence.NoOpMessagePersister;
 import io.joynr.provider.JoynrInterface;
 import io.joynr.provider.ProviderAnnotations;
 import io.joynr.runtime.AbstractJoynrApplication;
@@ -116,7 +113,6 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
 
     private RawMessagingPreprocessor rawMessagePreprocessor;
     private MqttClientIdProvider mqttClientIdProvider;
-    private MessagePersister messagePersister;
 
     /**
      * Constructor in which the JEE runtime injects the managed resources and the JEE joynr integration specific
@@ -132,7 +128,6 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
      * @param joynrLocalDomain the joynr local domain name to use for the application.
      * @param rawMessagePreprocessor can be optionally provided to intercept incoming messages and inspect or modify them
      * @param mqttClientIdProvider can be optionally provided to generate custom mqtt client id
-     * @param messagePersister can be optionally provided to persist joynr messages and thus increase sudden crash resilience
      * @param beanManager bean manager
      * @param joynrStatusMetrics Is passed to POJO joynr and receives metrics about the status of the mqtt connection.
      */
@@ -142,7 +137,6 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
                                       @JoynrLocalDomain Instance<String> joynrLocalDomain,
                                       @JoynrRawMessagingPreprocessor Instance<RawMessagingPreprocessor> rawMessagePreprocessor,
                                       @JoynrMqttClientIdProvider Instance<MqttClientIdProvider> mqttClientIdProvider,
-                                      @JoynrMessagePersister Instance<MessagePersister> messagePersister,
                                       BeanManager beanManager,
                                       JoynrStatusMetricsReceiver joynrStatusMetrics) {
         // CHECKSTYLE:ON
@@ -187,18 +181,6 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
             }
         } else {
             this.mqttClientIdProvider = null;
-        }
-
-        if (!messagePersister.isUnsatisfied()) {
-            if (!messagePersister.isAmbiguous()) {
-                this.messagePersister = messagePersister.get();
-            } else {
-                String message = "Only one MessagePersister may be provided.";
-                logger.error(message);
-                throw new JoynrIllegalStateException(message);
-            }
-        } else {
-            this.messagePersister = new NoOpMessagePersister();
         }
 
         Properties configuredProperties;
@@ -257,7 +239,6 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
                 @Override
                 protected void configure() {
                     bind(RawMessagingPreprocessor.class).toInstance(rawMessagePreprocessor);
-                    bind(MessagePersister.class).toInstance(messagePersister);
 
                     if (mqttClientIdProvider != null) {
                         bind(MqttClientIdProvider.class).toInstance(mqttClientIdProvider);
@@ -372,9 +353,5 @@ public class DefaultJoynrRuntimeFactory implements JoynrRuntimeFactory {
 
     public MqttClientIdProvider getMqttClientIdProvider() {
         return mqttClientIdProvider;
-    }
-
-    public MessagePersister getMessagePersister() {
-        return messagePersister;
     }
 }