+++ java/jeeintegration/src/test/java/io/joynr/jeeintegration/DefaultJoynrRuntimeFactoryTest.java
@@ -58,8 +58,6 @@ import io.joynr.messaging.MessagingQos;
 import io.joynr.messaging.NoOpRawMessagingPreprocessor;
 import io.joynr.messaging.RawMessagingPreprocessor;
 import io.joynr.messaging.mqtt.MqttClientIdProvider;
-import io.joynr.messaging.persistence.MessagePersister;
-import io.joynr.messaging.persistence.NoOpMessagePersister;
 import io.joynr.provider.ProviderAnnotations;
 import io.joynr.runtime.JoynrRuntime;
 import io.joynr.statusmetrics.JoynrStatusMetricsReceiver;
@@ -138,8 +136,6 @@ public class DefaultJoynrRuntimeFactoryTest {
                                Instance<String> joynrLocalDomain) throws Exception {
         Instance<RawMessagingPreprocessor> rawMessageProcessor = mock(Instance.class);
         when(rawMessageProcessor.get()).thenReturn(new NoOpRawMessagingPreprocessor());
-        Instance<MessagePersister> messagePersisterInstance = mock(Instance.class);
-        when(messagePersisterInstance.get()).thenReturn(new NoOpMessagePersister());
         BeanManager beanManager = mock(BeanManager.class);
         Bean<JoynrMessageProcessor> bean = mock(Bean.class);
         when(bean.create(Mockito.any())).thenReturn(new JoynrMessageProcessorTest());
@@ -158,7 +154,6 @@ public class DefaultJoynrRuntimeFactoryTest {
                                                  joynrLocalDomain,
                                                  rawMessageProcessor,
                                                  mqttClientIdProviderInstance,
-                                                 messagePersisterInstance,
                                                  beanManager,
                                                  joynrStatusMetrics);
         scheduledExecutorService = mock(ScheduledExecutorService.class);